add create_split_token.c
