+++ Makefile
@@ -7,6 +7,7 @@ SRCS	= \
 	srcs/create_cmd_path.c\
 	srcs/parser.c\
 	srcs/create_token_list.c\
+	srcs/create_split_token.c\
 	srcs/create_token_util.c\
 	srcs/utils/free_string_array.c\
 	srcs/utils/add_slash.c
